Page mapping: avoid TLB flush when creating a new mapping
There is no need to invalidate a page (and invoke a TLB shootdown
on all CPUs) when a new mapping is being created for that page.
TLBs only need to be flushed when an existing mapping is either
modified or deleted. TLB shootdowns are expensive on systems
with a high CPU count, and avoiding them fixes sluggishness issues
that have been observed when a user application starts and when
large memory allocations are made.
This commit changes map_with_complete() so that it invokes the
completion directly instead of passing it to a flush handler. Since
a side effect of this change is that multiple CPUs can request a
mapping in parallel as a result of a page fault (because the
pending fault created by one CPU is released synchronously, without
waiting for the other CPUs to flush their TLB), the code is being
changed so that no error occurs when a mapping is requested for an
already mapped virtual address.