x86-64 spinlock: avoid use of atomic operations when spinning
With this change, when a CPU is spinning to acquire a spinlock it
does not repeatedly execute atomic operations (which create traffic
on the system bus), but first checks whether the lock is free (with
a simple read from memory) before attempting to acquire it with an
atomic operation.
This change brings substantial performance improvements in VMs with
a high CPU count: throughput of a Go web server has been shown to
more than double (in terms of number of requests per second) in GCP
instances with 32 or 224 CPUs. Command line used for the test:
`ab -c100 -n100000 -k`.